Isolate lazy imports.
