+++ src/guppy/main.py
@@ -1,41 +1,22 @@
 """
-Main entry point for GuPPy (Guided Photometry Analysis in Python)
+Command-line entry point for GuPPy (Guided Photometry Analysis in Python).
+
+Keep this module cheap to import. The pipeline steps run their multiprocessing pools under
+the "spawn" start method, and a spawned worker is a fresh interpreter that must materialize
+``__main__`` before it can unpickle its task. Here ``__main__`` is the ``guppy`` console
+script, whose ``from guppy.main import main`` sits *above* its ``if __name__`` guard — so
+every worker executes that import, on every pool creation. Importing Panel and the page
+builders here therefore cost ~1.9s per pool instead of ~0.04s. The application itself lives
+in ``guppy.app``, imported below only once the CLI has decided to serve it.
 """
 
 import argparse
 
 from . import logging_config
 
-# Configured at import scope, not inside main(), so that every process which imports this
-# module gets handlers -- including the multiprocessing pool workers, which re-execute the
-# ``guppy`` console script under the "spawn" start method.
+# At import scope, not inside main(), so that every process importing this module gets
+# handlers -- including the spawned pool workers, whose logs would otherwise be discarded.
 logging_config.setup_logging()
-
-
-def serve_app(*, start_path: str | None = None) -> None:
-    """Serve the GuPPy application using Panel.
-
-    Serves the homepage plus the step result-view routes on one persistent server (each a
-    per-session factory). The views live on this never-torn-down server so the browser tab
-    is never abruptly disconnected. ``show=True`` opens the homepage at ``/``.
-    """
-    # Deliberately deferred rather than imported at module scope. The pipeline steps run
-    # multiprocessing pools under the "spawn" start method, and every spawned worker
-    # re-executes the ``guppy`` console script, which imports this module. Keeping Panel
-    # and the page builders out of import scope keeps that re-import at ~0.01s instead of
-    # ~1.5s per worker. Only ``logging_config`` stays above, so workers still get logging.
-    import panel as pn
-
-    from .orchestration.home import build_homepage
-    from .orchestration.preprocess_view import build_preprocess_view
-    from .orchestration.transients_view import build_transients_view
-
-    routes = {
-        "/": lambda: build_homepage(start_path=start_path),
-        "/preprocess-view": build_preprocess_view,
-        "/transients-view": build_transients_view,
-    }
-    pn.serve(routes, show=True)
 
 
 def main() -> None:
@@ -64,6 +45,9 @@ def main() -> None:
     if args.export_logs:
         logging_config.export_log_file()
         return
+
+    # Deferred so that merely importing this module stays cheap -- see the module docstring.
+    from .app import serve_app
 
     serve_app(start_path=args.start_path)
